feat: add Zernike modules for standard, fringe, and Noll indexing schemes with documentation
